[CORDA-1059]: Code review changes.
